# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/web/passage/add": {
            "post": {
                "description": "添加文章接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "类别"
                ],
                "summary": "添加文章",
                "parameters": [
                    {
                        "description": "文章标题",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "作者",
                        "name": "author",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "标签",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "点击数量",
                        "name": "look",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "类别",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "描述",
                        "name": "description",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "内容",
                        "name": "content",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "图片连接",
                        "name": "img_link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Passages"
                        }
                    }
                }
            }
        },
        "/web/passage/list": {
            "get": {
                "description": "描述信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "类别"
                ],
                "summary": "文章列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Passages": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "category": {
                    "description": "默认长度为255",
                    "type": "string"
                },
                "content": {
                    "description": "如果content字段为空或默认值，将不会返回改数据",
                    "type": "string"
                },
                "created_at": {
                    "description": "omitempty 前端可以不用传入该字段，自动生成默认值，返回值如果为空，则不会返回该字段。",
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "description": "primary_key定义主键   AUTO_INCREMENT定义自增",
                    "type": "integer"
                },
                "img_link": {
                    "description": "忽略这个字段",
                    "type": "string"
                },
                "look": {
                    "description": "如果look_num的值为0  则不会返回该字段",
                    "type": "integer"
                },
                "tag": {
                    "description": "类型为varchar  最大100 列名为 `tag`",
                    "type": "string"
                },
                "title": {
                    "description": "不为空，且唯一",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}